api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
asyncpg==0.29.0
httpx==0.25.2
nats-py==2.6.0
boto3==1.29.7
redis==5.0.1
python-multipart==0.0.6
python-dotenv==1.0.0